remove floating point
* now can run on x86-nemu without battles

+++ include/audio.h
@@ -51,7 +51,7 @@ VOID
 AUDIO_PlayMusic(
    INT       iNumRIX,
    BOOL      fLoop,
-   FLOAT     flFadeTime
+   INT       flFadeTime
 );
 
 BOOL
