+++ src/main.c
@@ -380,9 +380,9 @@ PAL_SplashScreen(
             {
                for (i = 0; i < 256; i++)
                {
-                  rgCurrentPalette[i].r = (BYTE)(palette[i].r * ((float)dwTime / 15000));
-                  rgCurrentPalette[i].g = (BYTE)(palette[i].g * ((float)dwTime / 15000));
-                  rgCurrentPalette[i].b = (BYTE)(palette[i].b * ((float)dwTime / 15000));
+                  rgCurrentPalette[i].r = (BYTE)(palette[i].r * dwTime / 15000);
+                  rgCurrentPalette[i].g = (BYTE)(palette[i].g * dwTime / 15000);
+                  rgCurrentPalette[i].b = (BYTE)(palette[i].b * dwTime / 15000);
                }
                VIDEO_SetPalette(rgCurrentPalette);
                VIDEO_UpdateSurfacePalette(lpBitmapDown);
